brain/requirements.txt
google-genai
watchdog
python-dotenv
python-json-logger
pydub>=0.25.1
